정수 > 실수 버그 수정

+++ lib/src/models/recipe_model.dart
@@ -27,7 +27,7 @@ class RecipeModel {
     subtitle = snapshot['subtitle'];
     content = snapshot['content'];
     cookingtime = snapshot['cookingtime'];
-    percent = snapshot['percent'];
+    percent = snapshot['percent'].toDouble();
     List<dynamic> comList = snapshot['comments'];
     commentCount = (comList != null) ? comList.length : 0 ;
     List<dynamic> favList = snapshot['favorits'];
